+++ autotransform/filter/factory.py
@@ -19,6 +19,7 @@ from typing import Any, Callable, Dict, Mapping
 from autotransform.config import fetcher as Config
 from autotransform.filter.base import Filter, FilterBundle
 from autotransform.filter.extension import ExtensionFilter
+from autotransform.filter.regex import RegexFilter
 from autotransform.filter.type import FilterType
 
 
@@ -38,6 +39,7 @@ class FilterFactory:
 
     _getters: Dict[FilterType, Callable[[bool, Mapping[str, Any]], Filter]] = {
         FilterType.EXTENSION: ExtensionFilter.from_data,
+        FilterType.REGEX: RegexFilter.from_data,
     }
 
     @staticmethod